datastores: refactor ProcessInfo to align with protobuf API
Refactor ProcessInfo struct to match protobuf naming and properly represent
containerized processes with separate namespace PIDs.

Changes:
- Rename EntityID (uint64) -> UniqueId (uint32) - matches hash output
- Rename PID/PPID -> HostPid/HostPpid - clarifies host namespace
- Add Pid/Ppid - namespace PIDs for containerized processes
- Add ParentUniqueId - efficient parent lookups
- Add ExitTime - process lifecycle tracking
- Update ProcessStore interface to use uint32 entityId consistently

Fixes dispatcher bug where both HostPid and Pid were set to ancestor.PID,
incorrectly representing containerized process ancestry.

+++ api/v1beta1/datastores/interfaces.go
@@ -43,17 +43,17 @@ type ProcessStore interface {
 
 	// GetProcess retrieves process information by entity ID
 	// Returns nil, false if the process is not found
-	GetProcess(entityId uint64) (*ProcessInfo, bool)
+	GetProcess(entityId uint32) (*ProcessInfo, bool)
 
 	// GetChildProcesses returns all child processes of the given process
 	// Returns empty slice if no children found
-	GetChildProcesses(entityId uint64) ([]*ProcessInfo, error)
+	GetChildProcesses(entityId uint32) ([]*ProcessInfo, error)
 
 	// GetAncestry retrieves the process ancestry chain up to maxDepth levels
 	// Returns slice of ProcessInfo with [0] = process itself, [1] = parent, [2] = grandparent, etc.
 	// If a parent is not found in the tree, the chain stops there
 	// Returns empty slice if maxDepth <= 0 or process not found
-	GetAncestry(entityId uint64, maxDepth int) ([]*ProcessInfo, error)
+	GetAncestry(entityId uint32, maxDepth int) ([]*ProcessInfo, error)
 }
 
 // ContainerStore provides access to container information
